# contracts/openapi/emcore-identity-access-api/v1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EMCORE Identity \u0026 Access API",
    "description": "Manages user registration, credential authentication, multi-factor authentication (MFA), step-up authorization workflows, JWT session token issuance, session revocation, workload service client identities, JWKS verification keys, and administrative user security status locking. Owns authentication and cryptographic tokens; does not own tenant role definitions or business permissions directly.\n\n**Service Ownership:** Identity \u0026 Access Security Team\n**Intended Consumers:** Platform clients, mobile apps, gateway middleware, and federated service consumers\n**Operational Mode:** Primarily synchronous HTTP API; critical domain mutations emit versioned outbox events and may operate under eventual consistency.\n**Confidentiality:** Public/Enterprise Consumer Contract.",
    "termsOfService": "https://emcore.platform/terms-of-service",
    "contact": {
      "name": "EMCORE Architecture \u0026 Support Team",
      "url": "https://emcore.platform/support",
      "email": "support@emcore.platform"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "EMCORE Development Standalone Host Server"
    },
    {
      "url": "https://api.emcore.platform/v1",
      "description": "EMCORE Platform Target Server"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "postApiV1AuthRegister",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "postApiV1AuthLogin",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          }
        }
      }
    },
    "/api/v1/auth/token/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "postApiV1AuthTokenRefresh",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "postApiV1AuthLogout",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "postApiV1AuthLogoutAll",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "getApiV1AuthSessions",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "operationId": "deleteApiV1AuthSessionsSessionId",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "description": "Unique per-request tracing identifier (ULID or UUID). Max length 64 chars. Propagated through middleware to backend logs.",
            "schema": {
              "type": "string",
              "default": "req_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "description": "Distributed transaction correlation identifier across multiple services. Preserved across outbox domain events and message boundaries.",
            "schema": {
              "type": "string",
              "default": "cor_01HPX7K7R5YZ2X90WY"
            }
          },
          {
            "name": "X-Client-Version",
            "in": "header",
            "description": "Calling consumer version string for diagnostics and feature deprecation telemetry.",
            "schema": {
              "type": "string",
              "default": "1.0.0-build2026"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Preferred locale for user-facing validation and problem details messaging (e.g., en-AE, ar-AE).",
            "schema": {
              "type": "string",
              "default": "en-AE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardSuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error (unhandled exception without Problem Details schema formatting)"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmcoreProblemDetails"
                },
                "example": {
                  "type": "https://emcore.platform/errors/403",
                  "title": "Forbidden",
                  "status": 403,
                  "code": "ACCESS_DENIED",
                  "detail": "Authenticated caller lacks sufficient tenant role or delegated permissions for this resource.",
                  "traceId": "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "description": "Current operational or workflow lifecycle state indicator.",
            "format": "int32",
            "nullable": true,
            "example": "Active"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "description": "RFC 7807 standardized problem details error payload."
      },
...
      "RegisterRequest": {
        "required": [
          "email",
          "mobile",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Normalized electronic mail communication address.",
            "example": "developer@emcore.platform"
          },
          "mobile": {
            "type": "string",
            "description": "International E.164 formatted mobile phone contact number.",
            "example": "\u002B971500000000"
          },
          "password": {
            "type": "string",
            "description": "Sensitive authentication secret credential. Must satisfy complex entropy requirements. Never logged or retained in plaintext.",
            "example": "********"
          }
        }
      },
      "RegisterResponse": {
        "required": [
          "userId",
          "email",
          "mobile",
          "status"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Opaque unique identity reference identifier (ULID/UUID format). Must not be treated as a sequential database key.",
            "example": "usr_01HPX7K7R5YZ2X90WY0001"
          },
          "email": {
            "type": "string",
            "description": "Normalized electronic mail communication address.",
            "example": "developer@emcore.platform"
          },
          "mobile": {
            "type": "string",
            "description": "International E.164 formatted mobile phone contact number.",
            "example": "\u002B971500000000"
          },
          "status": {
            "type": "string",
            "description": "Current operational or workflow lifecycle state indicator.",
            "example": "Active"
          },
          "message": {
            "type": "string",
            "default": "Registration successful. Please verify your email or mobile number."
          }
        }
      },
...
      "StandardSuccessResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "PRODUCTION GATEWAY JWT VALIDATION: NOT IMPLEMENTED. PRODUCTION FALLBACK TO TEST AUTH: PROHIBITED. DEVELOPMENT ONLY token verification. DEFERRED.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ClientCredentials": {
        "type": "oauth2",
        "description": "OAuth-style client credentials for service identities. Requires Client ID and Secret to obtain a short-lived bearer token from POST /api/v1/auth/token.",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://auth.emcore.platform/api/v1/auth/token",
            "scopes": {
              "api.access": "General API service access",
              "admin.write": "Elevated administrative state modifications"
            }
          }
        }
      },
      "StepUpToken": {
        "type": "apiKey",
        "description": "Identity Step-Up Flow: IMPLEMENTED. Downstream Step-Up Enforcement: NOT VERIFIED / NOT IMPLEMENTED.",
        "name": "X-StepUp-Token",
        "in": "header"
      },
      "WebhookHmac": {
        "type": "apiKey",
        "description": "PLANNED \u2014 NOT IMPLEMENTED.",
        "name": "X-Signature-256",
        "in": "header"
      }
    }
  },
...
}